Adds mac support to code-sync tweaks
- also, factored fallback sync into separate file, and default it to off
